Java-Spring-Boot-RESTful-Api/src/main/java/com/itau/carros/adapters/out/security/TokenService.java
package com.itau.carros.adapters.out.security;

import com.auth0.jwt.JWT;
import com.auth0.jwt.algorithms.Algorithm;
import com.auth0.jwt.exceptions.JWTVerificationException;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.stereotype.Service;

import java.time.Instant;
import java.time.LocalDateTime;
import java.time.ZoneOffset;

@Service
public class TokenService {

    @Value("${secret.key}")
    private String secret;


    public String gerarToken(String email) {
        var algoritmo = Algorithm.HMAC256(secret);
        return JWT.create()
                .withIssuer("CarAPI")
                .withSubject(email)
                .withExpiresAt(dataExpiracao())
                .sign(algoritmo);

    }

    public String validarToken(String tokenJWT) {
        try {
            var algoritmo = Algorithm.HMAC256(secret);
            return JWT.require(algoritmo)
                    .withIssuer("CarAPI")
                    .build()
                    .verify(tokenJWT)
                    .getSubject();

        } catch (JWTVerificationException exception) {
            throw new JWTVerificationException("Token inválido ou expirado!");
        }
    }

    private Instant dataExpiracao() {
        return LocalDateTime.now().plusHours(2).toInstant(ZoneOffset.of("-03:00"));
    }
}
